pyproject.toml
[tool.poetry]
name = "api-casa-dados"
version = "0.1.0"
description = ""
authors = ["havokz <matheusfeu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["uvicorn"], version = "^0.104.0"}
selenium = "^4.14.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
sqlalchemy = "^2.0.22"
undetected-chromedriver = "^3.5.3"
bcrypt = "^4.0.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic-settings = "^2.0.3"
aiosqlite = "^0.19.0"
pydantic = {extras = ["email"], version = "^2.4.2"}
python-multipart = "^0.0.6"
uvicorn = "^0.23.2"
validate-docbr = "^1.10.0"
numpy = {version = "^1.26.1", python = ">=3.9,<3.13"}
pandas = "^2.1.1"
openpyxl = "^3.1.2"
unidecode = "^1.3.7"
motor = "^3.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
